+++ NEWS.md
@@ -15,6 +15,8 @@
 - Fix bug to allow using polars without library(polars) (#355).
 - New methods `<LazyFrame>$optimization_toggle()` + `$profile()` and enable rust-polars feature
   CSE: "Activate common subplan elimination optimization" (#323)
+- Named expression e.g. `pl$select(newname = pl$lit(2))` as no longer experimental and allowed as
+  default.
   
 # polars 0.7.0